Fix: 도메인마다 컨트롤러 분리
